+++ lib/Braintree/Error/Codes.php
@@ -625,6 +625,7 @@ class Codes
     const TRANSACTION_REFUND_AMOUNT_IS_TOO_LARGE                                      = '91521';
     const TRANSACTION_REFUND_AUTH_HARD_DECLINED                                       = '915200';
     const TRANSACTION_REFUND_AUTH_SOFT_DECLINED                                       = '915201';
+    const TRANSACTION_SCA_EXEMPTION_INVALID                                           = '915213';
     const TRANSACTION_SERVICE_FEE_AMOUNT_CANNOT_BE_NEGATIVE                           = '91554';
     const TRANSACTION_SERVICE_FEE_AMOUNT_FORMAT_IS_INVALID                            = '91555';
     const TRANSACTION_SERVICE_FEE_AMOUNT_IS_TOO_LARGE                                 = '91556';